+++ GoRogue.Snippets/Program.cs
@@ -1,0 +1,6 @@
+﻿// See https://aka.ms/new-console-template for more information
+
+using GoRogue.Snippets.HowTos;
+
+Console.WriteLine("This project exists only to contain snippets for the documentation.");
+DiceNotation.ExampleCode();