+++ GoRogue/Factories/IFactoryObject.cs
@@ -4,7 +4,7 @@ namespace GoRogue.Factories
 {
     /// <summary>
     /// Interface that can optionally be implemented by objects created via a <see cref="Factory{TBlueprintID, TProduced}" /> or
-    /// <see cref="AdvancedFactory{TBlueprintID, TBlueprintConfig, TProduced}" />.  The <see cref="DefinitionId" /> property
+    /// <see cref="AdvancedFactory{TBlueprintID, TBlueprintConfig, TProduced}" />.  The <see cref="DefinitionID" /> property
     /// will be automatically set to the ID of the blueprint used to create the object when the factory's Create function is
     /// called.
     /// </summary>
@@ -15,6 +15,6 @@ namespace GoRogue.Factories
         /// The identifier of the blueprint that created this object. Do not set manually -- the factory
         /// will automatically set this field when the object is created.
         /// </summary>
-        TBlueprintID DefinitionId { get; set; }
+        TBlueprintID DefinitionID { get; set; }
     }
 }